requirements.txt
urllib3
requests
pyyaml
beautifulsoup4
